Surprise refactor of the tasks infrastructure, stubs for character classes.
Turns out the task infrastructure was lacking in refinement and appeared to be initializing tasks twice.  Now using a bit more tuple unpacking and dictionaries rather than lists and magic C-style array/list indexing.

Classes are officially stubbed.  The rewrite of task.py was basically to drive a better copy-paste job into a class loading system, which is essentially the same loading mechannic, but probably not enough to factor out and generalize.